tests/requirements-test.txt
# Test Dependencies for OwnLens Media Publishing Platform

# Testing Framework
pytest>=8.4.2
pytest-cov>=7.0.0
pytest-mock>=3.11.1
pytest-asyncio>=0.24.0
pytest-timeout>=2.3.1
pytest-xdist>=3.6.0  # Parallel test execution

# Mocking and Stubbing
responses>=0.25.0  # HTTP mocking
moto>=5.0.0  # AWS services mocking (for S3)
fakeredis>=2.25.0  # Redis mocking

# Test Data Generation
Faker>=28.0.0  # Generate fake test data
factory-boy>=3.3.0  # Test data factories

# Database Testing
pytest-postgresql>=5.0.0  # PostgreSQL test database
testcontainers>=4.9.0  # Docker containers for testing
  # Includes: PostgreSQL, Kafka, ClickHouse, Redis containers

# Spark Testing
chispa>=0.9.0  # Spark DataFrame assertions
pyspark-testing>=0.1.0  # Spark DataFrame testing utilities

# Code Quality (for tests)
pylint>=4.0.2
black>=25.9.0
mypy>=1.7.0

# Performance Testing
psutil>=5.9.0  # System and process utilities
memory-profiler>=0.61.0  # Memory profiling
py-spy>=0.3.14  # Sampling profiler (optional)

